Moved API client and endpoints for model objects into specialized subdirectory; update imports in components and routes
